requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
google-cloud-aiplatform
langchain-community

